// full-whm-exp/src/types.ts
export enum AppTab {
  DASHBOARD = 'DASHBOARD',
  OSINT_DASHBOARD = 'OSINT_DASHBOARD',
  EXTRACTOR = 'EXTRACTOR', 
  SQL_INJECT = 'SQL_INJECT', 
  CMS_EXPLOIT = 'CMS_EXPLOIT', 
  NETWORK = 'NETWORK', 
  BOTNET_CORE = 'BOTNET_CORE',
  KIMIKUKIU_TOOLS = 'KIMIKUKIU_TOOLS',
  WHOAMISEC_GPT = 'WHOAMISEC_GPT',
  MEDIA_CREATOR = 'MEDIA_CREATOR',
  SETTINGS = 'SETTINGS',
  QUANTUM_INTELLIGENCE = 'QUANTUM_INTELLIGENCE',
  ZXCDDOS = 'ZXCDDOS',
  IDE_TOOL = 'IDE_TOOL',
  ADMIN_PANEL = 'ADMIN_PANEL',
  DONATION_SUPPORT = 'DONATION_SUPPORT',
  GPT_TOOL = 'GPT_TOOL',
  SOLANA_TOOL = 'SOLANA_TOOL',
  DEPLOYER_TOOL = 'DEPLOYER_TOOL',
  QUANTUM_TOOL = 'QUANTUM_TOOL',
  SCANNER_TOOL = 'SCANNER_TOOL',
  S3_TOOL = 'S3_TOOL',
  BLACKHAT_TOOL = 'BLACKHAT_TOOL',
  LAZARUS_TOOL = 'LAZARUS_TOOL',
  BURPSUITE_TOOL = 'BURPSUITE_TOOL',
  OWASP_TOOL = 'OWASP_TOOL',
  OFFENSIVE_SECURITY = 'OFFENSIVE_SECURITY'
}